test(tui): split entry_row_line tests by rendering responsibility
`row_view_tests/entry_row_line.rs` reached 1040 lines, crossing
CLAUDE.md's warn threshold and appearing as the sole warn entry in
rinkaku's own report for this branch.

Extract the two independent render rules into sibling modules —
`risk_marker.rs` and `test_coverage_badge.rs` — rather than cutting the
file by line count. They are separate modules because the marker and
the badge are no longer coupled.

entry_row_line.rs: 1040 -> 722 lines.
